Fixed major issue in camp_removeMember
The "Remove Member" function of the camp didn't save the removed member
to disk. This allowed for an easy exploit where you could add a member,
take his gold/gear, remove the member and repeat. The party was only
saved to disk on entering the wilderness or leaving the game. Now the
character is saved to disk after being removed.